fix: Fix a bug in the lti_store filter.
The filter was using model_to_dict to serialize content from the
configuration object to share with the user of the filter.  However, the
django `model_to_dict` function does not serialize fields if they are
not editable.  So generated fields like the lti_1p3_public_jwk were not
getting serialized and shared.  This is an important field for the
xblock-lti-consumer repo where this filter is being consumed.

Also bump the version so we can do another release.